Create techtim.py
This is the "starter code" file with some functions implemented that I did the other day. We want to avoid using this using this, but we can use it as a guideline for what kind of a direction we want to go in moving forward. https://www.youtube.com/watch?v=uoR4ilCWwKA